animatedContainer: refactored: was created new files with pages and with provider         изменено:      lib/main.dart         новый файл:    lib/my_color_page.dart         новый файл:    lib/my_color_provider.dart         изменено:      pubspec.lock
